Group system v1
I've partially finished the full group system; at the moment, the user
is able to join the group and then post in the group. If they aren't a
member they can't post in the group and if they aren't admin, they can't
see the settings page and if they somehow manage to navigate to the
group settings page without admin permissions, the page will be set to
destroy the session and then require a non-existent file, thus
destroying the pages contents.