Better encryption
After reading about how terrible the SHA256 method of encrypting was, I
set out on a quest to change this using the password_hash function. The
operating system of the old server didn't support the version of PHP
(seemingly) that allowed it to be used. I changed the OS to an updated
version of what it was, spend a day reconfiguring everything and
reimporting things; now it is done and the password_hash function is
used, giving a random salt providing a random hash everything single
password.